initial commit
This is just a simple module to exec and view outputs on android from
termux. This module will provide a way to launch the `termux-share`
from proot evironment. There will be a server running on the actual
termux enviroment and a client inside the proot enviroment. The client
will signal the server to open the file.

And btw, termux-share is slow from the proot environment, that's why
this thing exists.